CFY-7821 Create manager id file in the usage collector scripts if it is missing
